[SPOT-305][REFACTOR] TokenProvider 인터페이스 도입
