+++ src/main/java/com/example/spot/auth/application/refactor/TokenReissueService.java
@@ -2,7 +2,7 @@ package com.example.spot.auth.application.refactor;
 
 import com.example.spot.auth.presentation.dto.token.TokenResponseDTO;
 
-public interface TokenService {
+public interface TokenReissueService {
 
 	// 리프레시 토큰을 사용하여 새로운 액세스 토큰을 발급
 	TokenResponseDTO.TokenDTO reissueToken(String refreshToken);
