+++ src/main/java/com/example/spot/common/security/utils/JwtTokenProvider.java
@@ -1,13 +1,22 @@
 package com.example.spot.common.security.utils;
 
-import com.example.spot.common.api.code.status.ErrorStatus;
-import com.example.spot.common.api.exception.GeneralException;
+import com.example.spot.auth.application.refactor.TokenProvider;
 import com.example.spot.auth.presentation.dto.token.TokenResponseDTO;
 import com.example.spot.auth.presentation.dto.token.TokenResponseDTO.TokenDTO;
-import io.jsonwebtoken.*;
+import com.example.spot.common.api.code.status.ErrorStatus;
+import com.example.spot.common.api.exception.GeneralException;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.UnsupportedJwtException;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import jakarta.servlet.http.HttpServletRequest;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
+import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -16,14 +25,10 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
-import java.util.Date;
-
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class JwtTokenProvider {
+public class JwtTokenProvider implements TokenProvider {
 
     @Value("${token.access_secret}")
     private String JWT_SECRET_KEY;
@@ -41,10 +46,11 @@ public class JwtTokenProvider {
 
     /**
      * 토큰을 생성합니다.
+     *
      * @param memberId 회원 ID
      * @return 생성된 토큰
      */
-    // 액세스 및 리프레시 토큰 생성
+    @Override
     public TokenDTO createToken(Long memberId) {
         // 현재 시간
         Date now = new Date();
@@ -53,18 +59,74 @@ public class JwtTokenProvider {
 
         // 토큰 DTO 반환
         return TokenDTO.builder()
-            .accessToken(accessToken)
-            .refreshToken(refreshToken)
-            .accessTokenExpiresIn(ACCESS_TOKEN_EXPIRATION_TIME)
-            .build();
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .accessTokenExpiresIn(ACCESS_TOKEN_EXPIRATION_TIME)
+                .build();
     }
 
-    //
+    /**
+     * 리프레시 토큰을 이용하여 토큰을 재발급합니다.
+     *
+     * @param refreshToken 리프레시 토큰
+     * @return 새로 발급된 토큰
+     */
+    @Override
+    public TokenDTO reissueToken(String refreshToken) {
+        Long memberId = getMemberIdByToken(refreshToken);
+        return createToken(memberId);
+    }
+
+    @Override
+    public boolean isTokenExpired(String token) {
+        return validateToken(token, true) == ErrorStatus._EXPIRED_JWT;
+    }
+
+    @Override
+    public boolean validateToken(String token) {
+        return validateToken(token, false) == null;
+    }
+
+    /**
+     * 토큰을 이용하여 사용자 인증을 수행합니다.
+     *
+     * @param token       토큰
+     * @param userDetails 사용자 정보
+     * @return 사용자 인증 정보
+     */
+    @Override
+    public Authentication getAuthentication(String token, UserDetails userDetails) {
+        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
+    }
+
+    /**
+     * 토큰을 해석하여 회원 ID를 반환합니다.
+     *
+     * @param request HTTP 요청
+     * @return 회원 ID
+     */
+    @Override
+    public String resolveToken(HttpServletRequest request) {
+        return resolveHeaderToken(request, "Authorization", "Bearer ");
+    }
+
+    /**
+     * 토큰을 이용하여 회원 ID를 반환합니다.
+     *
+     * @param token 토큰
+     * @return 회원 ID
+     */
+    @Override
+    public Long getMemberIdByToken(String token) {
+        Claims claims = getClaims(token);
+        return claims.get("memberId", Long.class);
+    }
 
     /**
      * 전화번호 인증을 위한 임시 토큰을 생성합니다
+     *
      * @param email 이메일
-     * @return  생성된 임시 토큰
+     * @return 생성된 임시 토큰
      */
     public TokenResponseDTO.TempTokenDTO createTempToken(String email) {
         Date now = new Date();
@@ -79,28 +141,28 @@ public class JwtTokenProvider {
 
     /**
      * JWT 토큰 생성 -> 위 createToken 메서드에서 호출
-     * @param memberId 회원 ID
-     * @param now 현재 시간
+     *
+     * @param memberId       회원 ID
+     * @param now            현재 시간
      * @param expirationTime 만료 시간
-     * @param tokenType 토큰 타입
+     * @param tokenType      토큰 타입
      * @return 생성된 토큰
      */
     private String generateToken(Long memberId, Date now, long expirationTime, String tokenType) {
         return Jwts.builder()
-            .claim("memberId", memberId) // 회원 ID
-            .claim("tokenType", tokenType) // 토큰 타입
-            .setIssuedAt(now) // 발급 시간
-            .setExpiration(new Date(now.getTime() + expirationTime)) // 만료 시간
-            .signWith(Keys.hmacShaKeyFor(JWT_SECRET_KEY.getBytes()), SignatureAlgorithm.HS256)
-            .compact();
+                .claim("memberId", memberId) // 회원 ID
+                .claim("tokenType", tokenType) // 토큰 타입
+                .setIssuedAt(now) // 발급 시간
+                .setExpiration(new Date(now.getTime() + expirationTime)) // 만료 시간
+                .signWith(Keys.hmacShaKeyFor(JWT_SECRET_KEY.getBytes()), SignatureAlgorithm.HS256)
+                .compact();
     }
-
-    // JWT 임시 토큰 생성 -> 위 createTempToken 메서드에서 호출
 
     /**
      * JWT 임시 토큰 생성
+     *
      * @param email 이메일
-     * @param now  현재 시간
+     * @param now   현재 시간
      * @return 생성된 임시 토큰
      */
     private String generateTempToken(String email, Date now) {
@@ -113,24 +175,17 @@ public class JwtTokenProvider {
                 .compact();
     }
 
-    // 토큰 유효성 검사 -> 유효기간 만료 여부 확인
-    public boolean isTokenExpired(String token) {
-        return validateToken(token, true) == ErrorStatus._EXPIRED_JWT;
-    }
 
-    // 토큰 유효성 검사 -> 외부 호출 용
-    public boolean validateToken(String token) {
-        return validateToken(token, false) == null;
-    }
-
-    // 토큰 유효성 검사
     private ErrorStatus validateToken(String token, boolean checkExpirationOnly) {
         try {
-            Jwts.parserBuilder().setSigningKey(Keys.hmacShaKeyFor(JWT_SECRET_KEY.getBytes())).build().parseClaimsJws(token);
+            Jwts.parserBuilder().setSigningKey(Keys.hmacShaKeyFor(JWT_SECRET_KEY.getBytes())).build()
+                    .parseClaimsJws(token);
             return null;
         } catch (ExpiredJwtException e) {
             // 만료된 토큰
-            if (checkExpirationOnly) return ErrorStatus._EXPIRED_JWT;
+            if (checkExpirationOnly) {
+                return ErrorStatus._EXPIRED_JWT;
+            }
             throw new GeneralException(ErrorStatus._EXPIRED_JWT);
         } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
             // 잘못된 JWT 서명
@@ -152,38 +207,11 @@ public class JwtTokenProvider {
     }
 
     /**
-     * 토큰을 이용하여 사용자 인증을 수행합니다.
-     * @param token 토큰
-     * @param userDetails 사용자 정보
-     * @return 사용자 인증 정보
-     */
-    public Authentication getAuthentication(String token, UserDetails userDetails) {
-        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
-    }
-
-    /**
-     * 토큰을 해석하여 회원 ID를 반환합니다.
-     * @param request HTTP 요청
-     * @return 회원 ID
-     */
-    public String resolveToken(HttpServletRequest request) {
-        return resolveHeaderToken(request, "Authorization", "Bearer ");
-    }
-
-    /**
-     * 리프레시 토큰을 해석하여 반환합니다.
-     * @param request HTTP 요청
-     * @return 리프레시 토큰
-     */
-    public String resolveRefreshToken(HttpServletRequest request) {
-        return resolveHeaderToken(request, "refreshToken", "");
-    }
-
-    /**
      * 헤더에서 토큰을 추출합니다.
-     * @param request HTTP 요청
+     *
+     * @param request    HTTP 요청
      * @param headerName 헤더 이름
-     * @param prefix 토큰 접두사
+     * @param prefix     토큰 접두사
      * @return 추출된 토큰
      */
     private String resolveHeaderToken(HttpServletRequest request, String headerName, String prefix) {
@@ -195,27 +223,8 @@ public class JwtTokenProvider {
     }
 
     /**
-     * 리프레시 토큰을 이용하여 토큰을 재발급합니다.
-     * @param refreshToken 리프레시 토큰
-     * @return 새로 발급된 토큰
-     */
-    public TokenDTO reissueToken(String refreshToken) {
-        Long memberId = getMemberIdByToken(refreshToken);
-        return createToken(memberId);
-    }
-
-    /**
-     * 토큰을 이용하여 회원 ID를 반환합니다.
-     * @param token 토큰
-     * @return 회원 ID
-     */
-    public Long getMemberIdByToken(String token) {
-        Claims claims = getClaims(token);
-        return claims.get("memberId", Long.class);
-    }
-
-    /**
      * 임시 토큰을 이용하여 이메일을 반환합니다.
+     *
      * @param tempToken 임시 토큰
      * @return 이메일
      */
@@ -226,6 +235,7 @@ public class JwtTokenProvider {
 
     /**
      * 토큰을 해석하여 클레임을 반환합니다.
+     *
      * @param token 토큰
      * @return 클레임
      */
